AESP-Microservice-Project/services/payment-service/requirements.txt
flask
flask-sqlalchemy
pymysql
pika
flask-cors
python-dotenv
requests
cryptography